requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
pydantic
browser-use
jinja2
python-multipart